+++ src/script/src/python.rs
@@ -14,19 +14,14 @@
 
 //! Python script coprocessor
 
-mod builtins;
-pub(crate) mod coprocessor;
-mod dataframe;
 mod engine;
 pub mod error;
-#[cfg(test)]
-mod test;
 pub(crate) mod utils;
-mod vector;
 
 pub use self::engine::{PyEngine, PyScript};
-pub use self::vector::PyVector;
 
 mod ffi_types;
 mod pyo3;
 mod rspython;
+
+pub(crate) use ffi_types::PyVector;
