+++ src/script/src/python/utils.rs
@@ -25,7 +25,7 @@ use rustpython_vm::builtins::{PyBaseExceptionRef, PyBool, PyFloat, PyInt, PyList
 use rustpython_vm::{PyObjectRef, PyPayload, PyRef, VirtualMachine};
 use snafu::{Backtrace, GenerateImplicitData, OptionExt, ResultExt};
 
-use crate::python::builtins::try_into_columnar_value;
+use crate::python::rspython::builtins::try_into_columnar_value;
 use crate::python::error::ret_other_error_with;
 use crate::python::{error, PyVector};
